Refactor : 주석 제거, 로직클래스와 게임클래스의 분리
