+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,7 @@
 package racingcar;
 
 public class Application {
-    private static final Logic logic = new Logic();
-
     public static void main(String[] args) {
-        Logic.doGame();
+        Game.doGame();
     }
 }
